requirements.txt
# SLRS — Requirements
# Ruaha Catholic University · BSCSE 2023

opencv-python==4.9.0.80
mediapipe==0.10.13
protobuf==4.25.8
tensorflow==2.16.2
keras>=3.0.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
matplotlib==3.8.3
seaborn==0.13.2
streamlit==1.32.2
Pillow==10.2.0
tqdm==4.66.2
pyttsx3==2.90
python-dotenv==1.0.1
